requirements.txt
 fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
newspaper3k
gitpython
python-multipart
httpx
apscheduler
